Rename Brain into _Brain
