+++ doc/python_reference.rst
@@ -268,7 +268,6 @@ Visualization
    set_3d_backend
    get_3d_backend
    use_3d_backend
-   Brain
 
 
 Preprocessing
