+++ mne/viz/__init__.py
@@ -26,4 +26,3 @@ from .ica import (plot_ica_scores, plot_ica_sources, plot_ica_overlay,
 from .montage import plot_montage
 from .backends.renderer import (set_3d_backend, get_3d_backend, use_3d_backend)
 from . import backends
-from .brain import Brain
